Updated types + fixed gradle file

+++ src/main/java/uk/gov/hmcts/reform/pip/model/CourtRoom.java
@@ -20,7 +20,7 @@ import javax.persistence.Id;
 public class CourtRoom {
 
     @Id
-    private int courtRoomId;
+    private Long courtRoomId;
 
     private Classification classification;
